+++ spec/cli_spec.rb
@@ -9,55 +9,62 @@ FAIL_APP_DIR = File.expand_path('test_apps/fail_app', __dir__)
 
 RSpec.describe 'CLI' do
   it 'renders reproducible dumps' do
-    cmd = %w[bundle exec rails-response-dumper]
-    stdout, stderr, status = Open3.capture3(*cmd, chdir: APP_DIR)
-    expect(stderr).to eq('')
-    expect(stdout).to eq("....\n")
-    expect(status.exitstatus).to eq(0)
+    dumps_dir = File.join(APP_DIR, 'dumps')
 
-    expect(File.join(APP_DIR, 'dumps')).to match_snapshots
+    begin
+      cmd = %w[bundle exec rails-response-dumper]
+      stdout, stderr, status = Open3.capture3(*cmd, chdir: APP_DIR)
+      expect(stderr).to eq('')
+      expect(stdout).to eq("....\n")
+      expect(status.exitstatus).to eq(0)
+
+      # defaults to 'dumps' directory
+      expect(dumps_dir).to match_snapshots
+    ensure
+      FileUtils.rm_rf dumps_dir
+    end
   end
 
   context 'with --verbose argument' do
     it 'outputs dumper and dump block names' do
-      cmd = %w[bundle exec rails-response-dumper --verbose]
+      cmd = %W[bundle exec rails-response-dumper --verbose --dumps-dir #{tmpdir}]
       stdout, stderr, status = Open3.capture3(*cmd, chdir: APP_DIR)
       expect(stderr).to eq('')
       expect(stdout).to eq("Hooks.hook .\nRoot.index .\nTests.post_with_body .\nTests.multiple_requests .\n\n")
       expect(status.exitstatus).to eq(0)
 
       # The snapshot output remains the same with the verbose option
-      expect(File.join(APP_DIR, 'dumps')).to match_snapshots
+      expect(tmpdir).to match_snapshots(File.join(APP_DIR, 'snapshots'))
     end
   end
 
   context 'with --exclude-response-headers argument' do
     it 'renders dumps without response headers' do
-      cmd = %w[bundle exec rails-response-dumper --exclude-response-headers]
+      cmd = %W[bundle exec rails-response-dumper --exclude-response-headers --dumps-dir #{tmpdir}]
       stdout, stderr, status = Open3.capture3(*cmd, chdir: APP_DIR)
       expect(stderr).to eq('')
       expect(stdout).to eq("....\n")
       expect(status.exitstatus).to eq(0)
 
-      expect(File.join(APP_DIR, 'dumps')).to match_snapshots(File.join(APP_DIR, 'snapshots_without_response_headers'))
+      expect(tmpdir).to match_snapshots(File.join(APP_DIR, 'snapshots_without_response_headers'))
     end
   end
 
   context 'when run with --order' do
     context 'with type "random"' do
       it 'renders reproducible dumps with random seed' do
-        cmd = %w[bundle exec rails-response-dumper --order random]
+        cmd = %W[bundle exec rails-response-dumper --order random --dumps-dir #{tmpdir}]
         stdout, stderr, status = Open3.capture3(*cmd, chdir: APP_DIR)
         expect(stderr).to eq('')
         expect(stdout).to match(/\ARandomized with seed [1-9][0-9]*\n\.\.\.\.\n\z/)
         expect(status.exitstatus).to eq(0)
-        expect(File.join(APP_DIR, 'dumps')).to match_snapshots
+        expect(tmpdir).to match_snapshots(File.join(APP_DIR, 'snapshots'))
       end
     end
 
     context 'when given a seed value' do
       it 'creates dump files in reproducible order' do
-        cmd = %w[bundle exec rails-response-dumper --order 8 --verbose]
+        cmd = %W[bundle exec rails-response-dumper --order 8 --verbose --dumps-dir #{tmpdir}]
         stdout, stderr, status = Open3.capture3(*cmd, chdir: APP_DIR)
         expect(stderr).to eq('')
         expect(stdout).to eq <<~TEXT
@@ -69,14 +76,14 @@ RSpec.describe 'CLI' do
 
         TEXT
         expect(status.exitstatus).to eq(0)
-        expect(File.join(APP_DIR, 'dumps')).to match_snapshots
+        expect(tmpdir).to match_snapshots(File.join(APP_DIR, 'snapshots'))
       end
     end
   end
 
   it 'runs after hook when an exception is raised' do
     env = { 'TMPDIR' => tmpdir, 'FILENAME' => 'test.out' }
-    cmd = %w[bundle exec rails-response-dumper]
+    cmd = %W[bundle exec rails-response-dumper --dumps-dir #{tmpdir}]
     stdout, stderr, status = Open3.capture3(env, *cmd, chdir: AFTER_HOOK_APP_DIR)
     expect(stderr).to eq('')
     expect(stdout.lines[0]).to eq("F\n")
@@ -93,7 +100,7 @@ RSpec.describe 'CLI' do
     let(:dumper_2_invalid_status_code) { 'fail_app_2.invalid_status_code' }
 
     it 'outputs all errors after execution' do
-      cmd = %w[bundle exec rails-response-dumper]
+      cmd = %W[bundle exec rails-response-dumper --dumps-dir #{tmpdir}]
       stdout, stderr, status = Open3.capture3(*cmd, chdir: FAIL_APP_DIR)
       expect(stderr).to eq('')
       expect(stdout.lines[0]).to eq("FFF\n")
@@ -114,7 +121,7 @@ RSpec.describe 'CLI' do
 
     context 'with --fail-fast argument' do
       it 'aborts after the first error' do
-        cmd = %w[bundle exec rails-response-dumper --fail-fast]
+        cmd = %W[bundle exec rails-response-dumper --fail-fast --dumps-dir #{tmpdir}]
         stdout, stderr, status = Open3.capture3(*cmd, chdir: FAIL_APP_DIR)
         expect(stderr).to eq('')
         expect(stdout.lines[0]).to eq("F\n")